divide navbar into 3 files
